backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary
requests==2.31.0
minio==7.2.0
PyJWT==2.8.0
